Move *.ncc to ncc/
Alex said that some of these are used, and others come from Noctis I and are
not. As a test, I moved them out and tried compiling, which the program did
successfully (both by build.bat and bc.exe).

I can think of a reason: these files aren't actually needed, because the code
already includes the models somehow.

The interior and exterior of the Stardrifter look fine, so I'll assume the
above is true. But if I notice a lack of fauna on Felysian planets, then I'll
be sure to put back the critters and dinner^H^H^H^H^H^H friendly avians. ;)